Move remaining listeners to extenders Drop toArray() settings feature that's not used
