+++ extend.php
@@ -4,7 +4,6 @@ namespace FoF\Upload;
 
 use Flarum\Extend;
 use Flarum\Foundation\Application;
-use Illuminate\Contracts\Events\Dispatcher;
 
 return [
     (new Extend\Routes('api'))
@@ -19,16 +18,15 @@ return [
         ->css(__DIR__ . '/resources/less/forum/upload.less')
         ->js(__DIR__ . '/js/dist/forum.js'),
     new Extend\Locales(__DIR__ . '/resources/locale'),
+    new Extenders\AddAvailableOptionsInAdmin(),
     new Extenders\AddForumAttributes(),
+    new Extenders\AddImageProcessor(),
     new Extenders\AddPostDownloadTags(),
     new Extenders\ReplaceDeprecatedTemplates(),
-    function (Dispatcher $events, Application $app) {
+    function (Application $app) {
         $app->register(Providers\SettingsProvider::class);
 
         $app->register(Providers\StorageServiceProvider::class);
         $app->register(Providers\DownloadProvider::class);
-
-        $events->subscribe(Listeners\LoadSettingsFromDatabase::class);
-        $events->subscribe(Listeners\ProcessesImages::class);
     },
 ];
